modify navbar
