--- react-app/src/components/NavBar.js
@@ -1,36 +1,0 @@
-
-import React from 'react';
-import { NavLink } from 'react-router-dom';
-import LogoutButton from './auth/LogoutButton';
-import LoginFormModal from './Splash/LoginFormModal';
-import SignupFormModal from './Splash/SignupFormModal';
-
-const NavBar = () => {
-  return (
-    <nav>
-      <ul>
-        <li>
-          <NavLink to='/' exact={true} activeClassName='active'>
-            Home
-          </NavLink>
-        </li>
-        <li>
-          <LoginFormModal />
-        </li>
-        <li>
-          <SignupFormModal />
-        </li>
-        <li>
-          <NavLink to='/users' exact={true} activeClassName='active'>
-            Users
-          </NavLink>
-        </li>
-        <li>
-          <LogoutButton />
-        </li>
-      </ul>
-    </nav>
-  );
-}
-
-export default NavBar;
