+++ react-app/src/components/Profile/index.js
@@ -1,0 +1,9 @@
+import './Profile.css'
+
+function Profile() {
+    return (
+        <h1>Profile Page!</h1>
+    );
+}
+
+export default Profile;
