+++ react-app/src/components/Splash/SignupFormModal/SignupForm.js
@@ -1,21 +1,24 @@
 import React, { useState } from 'react';
-import { useSelector, useDispatch } from 'react-redux';
+import { useSelector, useDispatch } from 'react-redux'
 // import { Redirect } from 'react-router-dom';
 import { signUp } from '../../../store/session';
 
-const SignupForm = () => {
+const SignUpForm = () => {
   const [errors, setErrors] = useState([]);
   const [username, setUsername] = useState('');
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
+  const [repeatPassword, setRepeatPassword] = useState('');
   // const user = useSelector(state => state.session.user);
   const dispatch = useDispatch();
 
-  const onLogin = async (e) => {
+  const onSignUp = async (e) => {
     e.preventDefault();
-    const data = await dispatch(signUp(username, email, password));
-    if (data) {
-      setErrors(data);
+    if (password === repeatPassword) {
+      const data = await dispatch(signUp(username, email, password));
+      if (data) {
+        setErrors(data)
+      }
     }
   };
 
@@ -31,50 +34,61 @@ const SignupForm = () => {
     setPassword(e.target.value);
   };
 
+  const updateRepeatPassword = (e) => {
+    setRepeatPassword(e.target.value);
+  };
+
   // if (user) {
   //   return <Redirect to='/' />;
   // }
 
   return (
-    <form onSubmit={onLogin}>
+    <form onSubmit={onSignUp}>
       <div>
         {errors.map((error, ind) => (
           <div key={ind}>{error}</div>
         ))}
       </div>
       <div>
-        <label htmlFor='username'>Username</label>
+        <label>User Name</label>
         <input
+          type='text'
           name='username'
-          type='text'
-          placeholder='Username'
-          value={username}
           onChange={updateUsername}
-        />
+          value={username}
+        ></input>
       </div>
       <div>
-        <label htmlFor='email'>Email</label>
+        <label>Email</label>
         <input
-          name='email'
           type='text'
-          placeholder='Email'
-          value={email}
+          name='email'
           onChange={updateEmail}
-        />
+          value={email}
+        ></input>
       </div>
       <div>
-        <label htmlFor='password'>Password</label>
+        <label>Password</label>
         <input
-          name='password'
           type='password'
-          placeholder='Password'
-          value={password}
+          name='password'
           onChange={updatePassword}
-        />
-        <button type='submit'>Sign Up</button>
+          value={password}
+        ></input>
       </div>
+      <div>
+        <label>Repeat Password</label>
+        <input
+          type='password'
+          name='repeat_password'
+          onChange={updateRepeatPassword}
+          value={repeatPassword}
+          required={true}
+        ></input>
+      </div>
+      <button type='submit'>Sign Up</button>
     </form>
   );
 };
 
-export default SignupForm;
+export default SignUpForm;
