+++ react-app/src/components/Splash/SignupFormModal/index.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { Modal } from '../../../context/Modal';
-import SignupForm from '../../auth/SignUpForm';
+import SignupForm from './SignUpForm';
 
 function SignupFormModal() {
   const [showModal, setShowModal] = useState(false);
