+++ react-app/src/components/Splash/SplashPage/index.js
@@ -1,14 +1,18 @@
-
-
+import React from 'react';
+import { useSelector } from 'react-redux';
+import { Redirect } from 'react-router';
 import './SplashPage.css'
 
 function SplashPage() {
+    const user = useSelector(state => state.session.user);
 
-
+    if (user) {
+      return <Redirect to={`/users/${user.id}`} />
+    };
 
     return (
-        <div className="SplashPage">
-            <h1>SPLASH!</h1>
+        <div className="splash-page-container">
+            <h1>Splash Page!</h1>
         </div>
     )
 }
